refactor: Move resolveTournaments function
